pyproject.toml
[project]
name = "enhanced-ai-system"
version = "0.1.0"
description = "Enhanced AI System with Multi-Model Integration"
requires-python = ">=3.9"

dependencies = [
    "aiohttp>=3.9.0",
    "asyncio>=3.4.3",
    "backoff>=2.0.0",
    "cachetools>=5.3.0",
    "networkx>=3.0",
    "numpy>=1.24.0", 
    "prometheus-client>=0.19.0",
    "psutil>=5.9.0",
    "pydantic>=2.5.0",
    "rich>=13.0.0",
    "tenacity>=8.0.0",
    "watchdog>=3.0.0"
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pyyaml>=6.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.hatch.build.targets.wheel]
packages = ["core"]
